+++ packages/lorena-sdk/package.json
@@ -1,16 +1,17 @@
 {
   "name": "@lorena-ssi/lorena-cli",
-  "version": "1.0.4",
+  "version": "1.0.5",
   "description": "Lorena Client",
   "author": "Alex Puig <alex@caelumlabs.com>",
-  "license": "ISC",
+  "license": "MIT",
   "main": "src/index.js",
+  "module": "src/main.js",
   "scripts": {
     "coverage": "nyc npm run mocha && nyc report --reporter=html",
     "coveralls": "nyc npm run mocha && nyc report --reporter=text-lcov | coveralls",
     "debug": "./pds 1",
     "lint": "eslint ./examples ./src ./test --fix",
-    "mocha": "mocha test --timeout 50000 --bail",
+    "mocha": "mocha test --timeout 50000 --bail --require esm",
     "test": "npm run mocha"
   },
   "husky": {
@@ -20,10 +21,9 @@
   },
   "dependencies": {
     "@lorena-ssi/matrix-lib": "^1.0.7",
-    "@lorena-ssi/zenroom-lib": "lorena-ssi/zenroom-lib#v1.3.3",
-    "axios": "^0.19.2",
-    "chalk-pipe": "^3.0.0",
+    "@lorena-ssi/zenroom-lib": "^1.3.3",
     "debug": "^4.1.1",
+    "esm": "^3.2.25",
     "inquirer": "^7.1.0"
   },
   "directories": {
@@ -38,15 +38,15 @@
     "eslint-config-standard": "^14.1.0",
     "eslint-plugin-chai-friendly": "^0.5.0",
     "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsdoc": "^18.4.3",
+    "eslint-plugin-jsdoc": "^22.0.1",
     "eslint-plugin-node": "^11.0.0",
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.1",
-    "husky": "^3.0.5",
-    "mocha": "^6.2.2",
+    "husky": "^4.2.3",
+    "mocha": "^7.1.0",
     "mocha-lcov-reporter": "^1.3.0",
     "nixt": "^0.5.1",
-    "nyc": "^14.1.1"
+    "nyc": "^15.0.0"
   },
   "repository": {
     "type": "git",
